UPD: changed unit tests

+++ lab_07_01_01/src/arr_handle.c
@@ -106,7 +106,7 @@ int key(const int *pb_src, const int *pe_src, int **pb_dst, int **pe_dst)
 
     *pe_dst = *pb_dst + dst_sz;
 
-    for (size_t i = 0; i < dst_sz; i++)
+    for (int i = 0; i < dst_sz; i++)
         *(*pb_dst + i) = *(pb_src + src_min_index + 1 + i);
 
     return EXIT_SUCCESS;
